micro-client/requirements.txt
grpcio-tools
etcd3==0.12.0
requests
